Разобьет утилиты на два файла
